requirements.txt
exceptiongroup==1.3.0
future==1.0.0
iniconfig==2.1.0
lxml==5.4.0
packaging==25.0
pluggy==1.5.0
pymavlink==2.4.47
pytest==8.3.5
PyYAML==6.0.2
tomli==2.2.1
typing_extensions==4.13.2
